Fix everything
- Everything should be working now.
- removed demo scene, added test scene
- zooming and move percentage now works.
- commented and formatted most of the code, might do some more later.
- changed XY and Rect structs to use double's instead of floats

+++ src/Structs.h
@@ -4,18 +4,19 @@
 
 namespace Structs
 {
+  // helpful structs for positions and rects
   struct XY
   {
-    float x = 0;
-    float y = 0;
+    double x = 0;
+    double y = 0;
   };
 
   struct Rect
   {
-    float x      = 0;
-    float y      = 0;
-    float width  = 0;
-    float height = 0;
+    double x      = 0;
+    double y      = 0;
+    double width  = 0;
+    double height  = 0;
   };
 }
